backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
requests
redis
PyPDF2
python-docx
pytest
httpx